add react-hot-loader, smaller margin in token modal heading
