+++ package.json
@@ -77,6 +77,7 @@
     "react": "^16.4.0",
     "react-dom": "^16.4.0",
     "react-final-form": "^4.1.0",
+    "react-hot-loader": "^4.8.0",
     "react-redux": "^6.0.1",
     "react-router-dom": "^4.3.1",
     "recompose": "^0.30.0",
